chore(scheduler-extension): adds helper to parse execution log (#745)
* chore(scheduler-extension): adds helper to parse execution log

scheduling a job, will execute a task on Windows, macOS, etc.
for this, use a file with pre-defined section

it'll help to know for which job-id it was, the time it took to
execute the job, the output of the job, etc.

related to https://github.com/kortex-hub/kortex/issues/519

Signed-off-by: Florent Benoit <fbenoit@redhat.com>